fix(agent): refresh VM reconcile lifecycle admission
